Fix resource leak when not able to connect
Properly handle interactions with gattlib to prevent resource leaks.

Fixes #4
