+++ examples/common.h
@@ -1,0 +1,7 @@
+#ifndef __COMMON_H__
+#define __COMMON_H__
+
+void *setup(size_t timeout, const char *address);
+int teardown(void);
+
+#endif /* __COMMON_H__ */
